Pull target to its own file
