+++ lib/cf_app_discovery.rb
@@ -7,6 +7,7 @@ require "cf_app_discovery/base"
 require "cf_app_discovery/auth"
 require "cf_app_discovery/client"
 require "cf_app_discovery/paginator"
+require "cf_app_discovery/target"
 require "cf_app_discovery/parser"
 require "cf_app_discovery/filestore_manager_factory"
 require "cf_app_discovery/target_configuration"
