+++ spec/cf_app_discovery/filter_spec.rb
@@ -3,7 +3,7 @@ require "spec_helper"
 RSpec.describe CfAppDiscovery::Filter do
   let(:targets) do
     [
-      CfAppDiscovery::Parser::Target.new(
+      CfAppDiscovery::Target.new(
         guid: "app-1-guid",
         name: "app-1",
         instances: 2,
@@ -13,7 +13,7 @@ RSpec.describe CfAppDiscovery::Filter do
         space: "test-space-name",
         org: "org-name"
       ),
-      CfAppDiscovery::Parser::Target.new(
+      CfAppDiscovery::Target.new(
         guid: "app-2-guid",
         name: "app-2",
         instances: 3,
@@ -23,7 +23,7 @@ RSpec.describe CfAppDiscovery::Filter do
         space: "test-space-name",
         org: "org-name"
       ),
-      CfAppDiscovery::Parser::Target.new(
+      CfAppDiscovery::Target.new(
         guid: "app-3-guid",
         name: "app-3",
         instances: 2,
